refactor: change models to repositories
